Moved QueryType Added QueryType to reconfig
