+++ starduck/src/adaptation/restart.rs
@@ -1,21 +1,7 @@
-use std::path::PathBuf;
-
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum QueryType {
-    Http { endpoint: PathBuf, port: u16 },
-}
-
-impl QueryType {
-    pub fn get_endpoint(&self) -> Option<PathBuf> {
-        match self {
-            QueryType::Http { endpoint, .. } => Some(endpoint.clone()),
-            _ => None,
-        }
-    }
-}
+use super::QueryType;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct RestartOrder {
